refactor: extraer alert en un componente aparte
