+++ src/components/AdvertenciaModal.tsx
@@ -1,0 +1,57 @@
+import {
+  AlertDialog,
+  AlertDialogOverlay,
+  AlertDialogContent,
+  AlertDialogHeader,
+  AlertDialogBody,
+  AlertDialogFooter,
+} from '@chakra-ui/modal'
+import { Button } from '@chakra-ui/react'
+import { useRef } from 'react'
+
+const AdvertenciaModal = ({
+  isOpen,
+  onClose,
+  onDelete,
+  alertHeader,
+  alertBody,
+}: any) => {
+  const cancelRef = useRef(null)
+
+  return (
+    <AlertDialog
+      isOpen={isOpen}
+      leastDestructiveRef={cancelRef}
+      onClose={onClose}
+      isCentered
+    >
+      <AlertDialogOverlay>
+        <AlertDialogContent>
+          <AlertDialogHeader fontSize="lg" fontWeight="bold">
+            {alertHeader}
+          </AlertDialogHeader>
+
+          <AlertDialogBody>{alertBody}</AlertDialogBody>
+
+          <AlertDialogFooter>
+            <Button ref={cancelRef} onClick={onClose}>
+              Cancelar
+            </Button>
+            <Button
+              colorScheme="red"
+              onClick={() => {
+                onDelete()
+                onClose()
+              }}
+              ml={3}
+            >
+              Borrar
+            </Button>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialogOverlay>
+    </AlertDialog>
+  )
+}
+
+export default AdvertenciaModal
